improvement: :sparkles: update profile update with consent x9
